[SIDE-2331] Rename/move mappers to aptitude/mapper.ts
